feat: int to float

+++ src/money/Expression.go
@@ -3,5 +3,5 @@ package money
 type Expression interface {
 	Reduce(Bank, string) Money
 	Plus(Expression) Expression
-	Times(int) Expression
+	Times(float64) Expression
 }
